// src/infrastructure/grpc/generated/course/types/stats.ts
export interface GetInstructorCourseRatingStatsRequest {
  courseId: string;
  instructorId: string;
}

export interface GetInstructorCourseRevenueSummeryRequest {
  courseId: string;
  instructorId: string;
}

export interface RatingsBreakDown {
  "1": number;
  "2": number;
  "3": number;
  "4": number;
  "5": number;
}

export interface InstructorCourseRatingStats {
  averageRating: number;
  totalRatings: number;
  breakdown: RatingsBreakDown | undefined;
}
// ...
export const GetInstructorCourseRatingStatsRequest: MessageFns<GetInstructorCourseRatingStatsRequest> =
  {
    encode(
      message: GetInstructorCourseRatingStatsRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.courseId !== "") {
        writer.uint32(10).string(message.courseId);
      }
      if (message.instructorId !== "") {
        writer.uint32(18).string(message.instructorId);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetInstructorCourseRatingStatsRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetInstructorCourseRatingStatsRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.courseId = reader.string();
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.instructorId = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetInstructorCourseRatingStatsRequest {
      return {
        courseId: isSet(object.courseId)
          ? globalThis.String(object.courseId)
          : isSet(object.course_id)
            ? globalThis.String(object.course_id)
            : "",
        instructorId: isSet(object.instructorId)
          ? globalThis.String(object.instructorId)
          : isSet(object.instructor_id)
            ? globalThis.String(object.instructor_id)
            : "",
      };
    },

    toJSON(message: GetInstructorCourseRatingStatsRequest): unknown {
      const obj: any = {};
      if (message.courseId !== "") {
        obj.courseId = message.courseId;
      }
      if (message.instructorId !== "") {
        obj.instructorId = message.instructorId;
      }
      return obj;
    },

    create<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsRequest>, I>,
    >(base?: I): GetInstructorCourseRatingStatsRequest {
      return GetInstructorCourseRatingStatsRequest.fromPartial(
        base ?? ({} as any),
      );
    },
    fromPartial<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsRequest>, I>,
    >(object: I): GetInstructorCourseRatingStatsRequest {
      const message = createBaseGetInstructorCourseRatingStatsRequest();
      message.courseId = object.courseId ?? "";
      message.instructorId = object.instructorId ?? "";
      return message;
    },
  };
// ...
export const RatingsBreakDown: MessageFns<RatingsBreakDown> = {
  encode(
    message: RatingsBreakDown,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message["1"] !== 0) {
      writer.uint32(8).int32(message["1"]);
    }
    if (message["2"] !== 0) {
      writer.uint32(16).int32(message["2"]);
    }
    if (message["3"] !== 0) {
      writer.uint32(24).int32(message["3"]);
    }
    if (message["4"] !== 0) {
      writer.uint32(32).int32(message["4"]);
    }
    if (message["5"] !== 0) {
      writer.uint32(40).int32(message["5"]);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RatingsBreakDown {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRatingsBreakDown();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message["1"] = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message["2"] = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message["3"] = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message["4"] = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message["5"] = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): RatingsBreakDown {
    return {
      "1": isSet(object["1"])
        ? globalThis.Number(object["1"])
        : isSet(object._1)
          ? globalThis.Number(object._1)
          : 0,
      "2": isSet(object["2"])
        ? globalThis.Number(object["2"])
        : isSet(object._2)
          ? globalThis.Number(object._2)
          : 0,
      "3": isSet(object["3"])
        ? globalThis.Number(object["3"])
        : isSet(object._3)
          ? globalThis.Number(object._3)
          : 0,
      "4": isSet(object["4"])
        ? globalThis.Number(object["4"])
        : isSet(object._4)
          ? globalThis.Number(object._4)
          : 0,
      "5": isSet(object["5"])
        ? globalThis.Number(object["5"])
        : isSet(object._5)
          ? globalThis.Number(object._5)
          : 0,
    };
  },

  toJSON(message: RatingsBreakDown): unknown {
    const obj: any = {};
    if (message["1"] !== 0) {
      obj["1"] = Math.round(message["1"]);
    }
    if (message["2"] !== 0) {
      obj["2"] = Math.round(message["2"]);
    }
    if (message["3"] !== 0) {
      obj["3"] = Math.round(message["3"]);
    }
    if (message["4"] !== 0) {
      obj["4"] = Math.round(message["4"]);
    }
    if (message["5"] !== 0) {
      obj["5"] = Math.round(message["5"]);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<RatingsBreakDown>, I>>(
    base?: I,
  ): RatingsBreakDown {
    return RatingsBreakDown.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RatingsBreakDown>, I>>(
    object: I,
  ): RatingsBreakDown {
    const message = createBaseRatingsBreakDown();
    message["1"] = object["1"] ?? 0;
    message["2"] = object["2"] ?? 0;
    message["3"] = object["3"] ?? 0;
    message["4"] = object["4"] ?? 0;
    message["5"] = object["5"] ?? 0;
    return message;
  },
};

function createBaseInstructorCourseRatingStats(): InstructorCourseRatingStats {
  return { averageRating: 0, totalRatings: 0, breakdown: undefined };
}

export const InstructorCourseRatingStats: MessageFns<InstructorCourseRatingStats> =
  {
    encode(
      message: InstructorCourseRatingStats,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.averageRating !== 0) {
        writer.uint32(9).double(message.averageRating);
      }
      if (message.totalRatings !== 0) {
        writer.uint32(17).double(message.totalRatings);
      }
      if (message.breakdown !== undefined) {
        RatingsBreakDown.encode(
          message.breakdown,
          writer.uint32(26).fork(),
        ).join();
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): InstructorCourseRatingStats {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseInstructorCourseRatingStats();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 9) {
              break;
            }

            message.averageRating = reader.double();
            continue;
          }
          case 2: {
            if (tag !== 17) {
              break;
            }

            message.totalRatings = reader.double();
            continue;
          }
          case 3: {
            if (tag !== 26) {
              break;
            }

            message.breakdown = RatingsBreakDown.decode(
              reader,
              reader.uint32(),
            );
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): InstructorCourseRatingStats {
      return {
        averageRating: isSet(object.averageRating)
          ? globalThis.Number(object.averageRating)
          : isSet(object.average_rating)
            ? globalThis.Number(object.average_rating)
            : 0,
        totalRatings: isSet(object.totalRatings)
          ? globalThis.Number(object.totalRatings)
          : isSet(object.total_ratings)
            ? globalThis.Number(object.total_ratings)
            : 0,
        breakdown: isSet(object.breakdown)
          ? RatingsBreakDown.fromJSON(object.breakdown)
          : undefined,
      };
    },

    toJSON(message: InstructorCourseRatingStats): unknown {
      const obj: any = {};
      if (message.averageRating !== 0) {
        obj.averageRating = message.averageRating;
      }
      if (message.totalRatings !== 0) {
        obj.totalRatings = message.totalRatings;
      }
      if (message.breakdown !== undefined) {
        obj.breakdown = RatingsBreakDown.toJSON(message.breakdown);
      }
      return obj;
    },

    create<I extends Exact<DeepPartial<InstructorCourseRatingStats>, I>>(
      base?: I,
    ): InstructorCourseRatingStats {
      return InstructorCourseRatingStats.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<InstructorCourseRatingStats>, I>>(
      object: I,
    ): InstructorCourseRatingStats {
      const message = createBaseInstructorCourseRatingStats();
      message.averageRating = object.averageRating ?? 0;
      message.totalRatings = object.totalRatings ?? 0;
      message.breakdown =
        object.breakdown !== undefined && object.breakdown !== null
          ? RatingsBreakDown.fromPartial(object.breakdown)
          : undefined;
      return message;
    },
  };
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}